consolidate test utils
